move test functions that differ from other ciphersuites to make gencode simpler to use (#364)
