+++ frost-ristretto255/tests/frost.rs
@@ -1,5 +1,3 @@
-use curve25519_dalek::{ristretto::RistrettoPoint, traits::Identity};
-use frost_core::{Ciphersuite, Group, GroupError};
 use frost_ristretto255::*;
 use lazy_static::lazy_static;
 use rand::thread_rng;
@@ -31,14 +29,6 @@ fn check_bad_batch_verify() {
     let rng = thread_rng();
 
     frost_core::tests::batch::bad_batch_verify::<Ristretto255Sha512, _>(rng);
-}
-
-#[test]
-fn check_deserialize_identity() {
-    let encoded_identity = RistrettoPoint::identity().compress().to_bytes();
-
-    let r = <Ristretto255Sha512 as Ciphersuite>::Group::deserialize(&encoded_identity);
-    assert_eq!(r, Err(GroupError::InvalidIdentityElement));
 }
 
 lazy_static! {
